Refactor unit summaries into ui-components package
