+++ apps/public/pages/listings.tsx
@@ -6,7 +6,6 @@ import Layout from "../layouts/application"
 import PageHeader from "@bloom-housing/ui-components/src/headers/page_header/page_header"
 import ListingsList from "@bloom-housing/ui-components/src/page_components/listing/ListingsList"
 import axios from "axios"
-import { groupNonReservedAndReservedSummaries } from "../lib/tableSummaries"
 import moment from "moment"
 import { Listing } from "@bloom-housing/core/src/listings"
 import ListingsGroup from "@bloom-housing/ui-components/src/page_components/listing/ListingsGroup"
@@ -38,10 +37,7 @@ export default class extends Component<ListingsProps> {
 
   renderOpenListings() {
     return this.props.openListings.length > 0 ? (
-      <ListingsList
-        listings={this.props.openListings}
-        unitSummariesTable={groupNonReservedAndReservedSummaries}
-      />
+      <ListingsList listings={this.props.openListings} />
     ) : (
       <div className="flex flex-row flex-wrap max-w-5xl m-auto mt-5 mb-12 text-center p-4 bg-gray-300">
         <h3 className="m-auto text-gray-800">{t("listings.noOpenListings")}</h3>
@@ -57,7 +53,6 @@ export default class extends Component<ListingsProps> {
           header={t("listings.closedListings")}
           hideButtonText={t("listings.hideClosedListings")}
           showButtonText={t("listings.showClosedListings")}
-          unitSummariesTable={groupNonReservedAndReservedSummaries}
         />
       )
     )
