+++ shared/ui-components/src/helpers/occupancyFormatting.tsx
@@ -1,0 +1,39 @@
+import * as React from "react"
+import t from "./translator"
+import { Listing } from "@bloom-housing/core/src/listings"
+
+export const occupancyTable = (listing: Listing) => {
+  const occupancyData = listing.unitsSummarized.byUnitType.map(unitSummary => {
+    let occupancy = ""
+
+    if (unitSummary.occupancyRange.max == null) {
+      occupancy = `at least ${unitSummary.occupancyRange.min} ${
+        unitSummary.occupancyRange.min == 1 ? t("listings.person") : t("listings.people")
+      }`
+    } else if (unitSummary.occupancyRange.max > 1) {
+      occupancy = `${unitSummary.occupancyRange.min}-${unitSummary.occupancyRange.max} ${
+        unitSummary.occupancyRange.max == 1 ? t("listings.person") : t("listings.people")
+      }`
+    } else {
+      occupancy = `1 ${t("listings.person")}`
+    }
+
+    return {
+      unitType: <strong>{t("listings.unitTypes." + unitSummary.unitType)}</strong>,
+      occupancy: occupancy
+    }
+  })
+
+  return occupancyData
+}
+
+export const getOccupancyDescription = (listing: Listing) => {
+  const unitsSummarized = listing.unitsSummarized
+  if (unitsSummarized.unitTypes.includes("SRO")) {
+    return unitsSummarized.unitTypes.length == 1
+      ? t("listings.occupancyDescriptionAllSro")
+      : t("listings.occupancyDescriptionSomeSro")
+  } else {
+    return t("listings.occupancyDescriptionNoSro")
+  }
+}
